Changed handling of return values for getList() and getCollection().
If the passed in key cannot be resolved and the default value is null, the
methods now return null rather than an empty collection. This is the
behavior which was used by DataConfiguration in version 1.x, and it makes
sense to return the default value directly if no data is found.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1516577 13f79535-47bb-0310-9956-ffa450edef68