Added a special handling for empty elements.
When converting to a list or to an array it makes sense to handle empty
elements in a configuration in a special way - rather than returning a
collection with a single empty element, an empty collection should be
returned. This is now implemented by DefaultConversionHandler. A new method was
added which determines whether an element is empty. So this can be customized
by subclasses.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1516578 13f79535-47bb-0310-9956-ffa450edef68