Added a method for checking whether a locator is fully initialized.
Because the way a file is referenced by a FileLocator can make a difference
when the file is actually accessed it makes sense to have a method which checks
this. This helps client code to decide whether fullyInitializedLocator()
needs to be called. FileHandler also needs this information to provide
functionality to initialize its FileLocator on demand.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1524819 13f79535-47bb-0310-9956-ffa450edef68