[CONFIGURATION-555] Fixed a problem with the handling of the xml:space attribute.
The attribute was only evaluated for sub elements, but not for the current
element. However, now there is a corner case of an element which only
contains sub elements and has the attribute set to "preserve". In this case,
a trim is done because it does not make sense to assign a value consisting
only of whitespace to this element.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1527396 13f79535-47bb-0310-9956-ffa450edef68