[CONFIGURATION-557] Regression: MapConfiguration no longer accepts a Map<String, String>
In 1.7 and before, it was possible to pass an arbitrary Map into the
constructor of MapConfiguration. With the generification in 1.8, this
actually broke and it was no longer possible to pass in e.g. a
Map<String, String> because the signature now required a Map<String,
Object>. Changing the constructor to accept a Map<String, ?> restores
this.

All of this is purely a compiler issue, the runtime itself does not
see any of the generics due to the Java type erasure.



git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/branches/CONFIGURATION_1_X_MAINTENANCE@1534048 13f79535-47bb-0310-9956-ffa450edef68