NodeModel no longer defines the getRootNode() method.
The root node now has to be obtained from the NodeHandler. This is necessary at
least for implementations like InMemoryNodeModel which use immutable structures
that can be replaced by new instances. A NodeHandler is a kind of snapshot of
such a structure at a certain point in time. If the root node is obtained in a
different method call, it may no longer be compatible with the snapshot used by
the node handler.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/branches/immutableNodes@1575763 13f79535-47bb-0310-9956-ffa450edef68