InMemoryModel now supports clearProperty() on a tracked node.
When executing update operations on tracked nodes it has to be distinguished
whether the node is detached or not. For Non-detached nodes the update
operation can be performed on the model's current node structure. Detached
nodes live in an independent node structure. This is implemented by storing
them in a separate node model.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/branches/immutableNodes@1575764 13f79535-47bb-0310-9956-ffa450edef68