Deprecated methods for handling configuration listeners.
The goal is that BasicConfigurationBuilder only provides the methods from the
EventSource interface for handling event listeners. This makes the API more
stringent and reduces confusion. Because of the filtering by event types it is
no problem to add listeners to managed configurations which they do not
support.

Currently some tests are failing and have been @Ignored.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1612679 13f79535-47bb-0310-9956-ffa450edef68