[CONFIGURATION-604] Changed construction of hierarchical configuration.
The conversion of a flat configuration to a hierarchical one now uses copy()
rather than append(). This causes a slightly different structure of the
resulting configuration. Because append() adds new nodes on each invocation
the property keys "x.y.z" and "x.y" processed in this order generates two y
child nodes of x. With copy() only a single node is created. This is closer to
the original structure of the flat configuration and is less dependent on the
order of configuration keys.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1690811 13f79535-47bb-0310-9956-ffa450edef68