[CONFIGURATION-646] Changed handling of loads in layout.
PropertiesConfigurationLayout used to remove its event listener
registration when a load operation starts and add it again
afterwards. This caused problems with the auto-save mechanism
because update events were processed by the layout after the listener
for auto-save was triggered, resulting in incomplete configuration
files.

Now the listener is not removed, but temporarily disabled.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1775739 13f79535-47bb-0310-9956-ffa450edef68