[CONFIGURATION-649] Improved list handling in XMLConfiguration.
List properties defined as single string with a delimiter character
now retain their format when the configuration is saved.

Note that the implemented solution is not perfect, it cannot handle
all possible constellations. For instance, for lists defined in a
mixed form (multiple XML elements each of wich defines multiple
list values) cannot be handled and are reformatted on saving.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1779753 13f79535-47bb-0310-9956-ffa450edef68