Fail build on checkstyle validation
Commons-configuration already has a vast set of checkstyle rules
configured, and most of the code adheres to them. However, they are
just reported, not enforced, so a developer working on a patch needs
to go over mvn's output himself and visually check for any checkstyle
violation reports.

This patch sets the failOnViolation property to true so violating
checkstyle's rules actually fails the build and forces the developer
to address any issue he may have introduced, even if it's just by
suppressing it.