[CONFIGURATION-712] Added a refresh() method to reloading detector.
FileHandlerReloadingDetector had the problem that the first invocation
of the isReloadingRequired() method the file's last-modified date was
initialized. So a changed file could only be recognized on a 2nd
invocation.

Now the class offers a dedicated method to initialize the
last-modified date from the underlying file.
DefaultReloadingDetectorFactory was changed to call this method after
the detector has been created. So reloading checks are active from the
very beginning.

Thanks to Rolland Hobbie for the patch.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/configuration/trunk@1841033 13f79535-47bb-0310-9956-ffa450edef68