Guard DefaultExpressionEngineSymbols against empty builder
If the symbols provided to the `DefaultExpressionEngineSymbols` is
`null`, then they are saved to the respective `private final`
properties, which is bound to throw an NPE at a later time. One
prominent case of this is when the builder is not modified before
creating an instance. This commit implements a `null`-guard to
use the default values when `null` is provided.